Move IUpdatableTable to right package
